Added Upgrade System & Improved Spawn System
Added an upgrade system that is relatively modular as it allows for most upgrades to be done in a small amount of time by simply just referencing the name of the upgrade in the script.
The wave spawn system was improved via a method of calculating how often and how many enemies should via an equation.